requirements.txt
ultralytics
fastapi
uvicorn
python-multipart 
Pillow
numpy<2
opencv-python-headless
